requirements.txt
# librerias de python
streamlit
google-generativeai
python-dotenv